Navbar search funcionando
