+++ src/styles/base/_settings.scss
@@ -1,0 +1,5 @@
+// Colors
+
+$primary: #5c62c5;
+$dark-gray: #363636;
+$light-gray: #d8d8d8;
