+++ src/styles/components/Navbar.scss
@@ -1,0 +1,3 @@
+.fa-film {
+  font-size: 4rem;
+}
