+++ src/styles/styles.scss
@@ -1,0 +1,6 @@
+@import './base/settings';
+@import './base/base';
+
+// Components
+@import './components/Navbar.scss';
+// @import './components/auth';
